Abstracting database
Just creating an interface of database so we can change it to use another implementation such as JPA, Hibernate without changing the consuming code
